requirements.txt
openai-whisper
pandas
pytest
